converts declearn folder to declearn.py
