+++ fedbiomed/common/optimizers/__init__.py
@@ -19,4 +19,4 @@ __all__ = [
     "NativeTorchOptimizer",
     "BaseOptimizer",
     "DeclearnOptimizer"
-           ]
+]
