--- fedbiomed/common/optimizers/declearn/__init__.py
@@ -1,7 +1,0 @@
-# This file is originally part of Fed-BioMed
-# SPDX-License-Identifier: Apache-2.0
-
-from .declearn_optimizer import list_optim_modules
-from .declearn_regularizer import list_optim_regularizers
-from .declearn_optimizer import *
-from .declearn_regularizer import *
